chore: more feedback and disabled inputs from cursor
