+++ src/ui/shared/index.ts
@@ -32,3 +32,4 @@ export * from "./icons";
 export * from "./box";
 export * from "./pre-code";
 export * from "./error-boundary";
+export * from "./how-it-works-footer";
